Rename file firebase => firestore
